Revert "Merge remote-tracking branch 'upstream/main' into v-danurg/pester_tests_fail_on_error"
This reverts commit fc70c7c37cc682a217ded02b1e79855ed2146602, reversing
changes made to 8feb503f7efeb462b584a15d887157f67c84af4d.

+++ images/linux/scripts/installers/selenium.sh
@@ -4,7 +4,6 @@
 ##  Desc:  Installs selenium server
 ################################################################################
 
-
 # Determine latest selenium standalone server version
 SELENIUM_LATEST_VERSION_URL=https://api.github.com/repos/SeleniumHQ/selenium/releases/latest
 SELENIUM_VERSION=$(curl $SELENIUM_LATEST_VERSION_URL | jq '.name' | tr -d '"' | cut -d ' ' -f 2)
@@ -15,8 +14,12 @@ echo "Downloading selenium-server-standalone v$SELENIUM_VERSION..."
 SELENIUM_JAR_NAME="selenium-server-standalone-$SELENIUM_VERSION.jar"
 wget https://selenium-release.storage.googleapis.com/$SELENIUM_VERSION_MAJOR_MINOR/$SELENIUM_JAR_NAME
 
+echo "Testing to make sure that script performed as expected, and basic scenarios work"
+if [ ! -f "$SELENIUM_JAR_NAME" ]; then
+    echo "Selenium server was not installed"
+    exit 1
+fi
+
 SELENIUM_JAR_PATH="/usr/share/java/selenium-server-standalone.jar"
 mv $SELENIUM_JAR_NAME $SELENIUM_JAR_PATH
 echo "SELENIUM_JAR_PATH=$SELENIUM_JAR_PATH" | tee -a /etc/environment
-
-invoke_tests "Tools" "Selenium"
